Fix memory corruption bug causing lines in output image + refactorings
The fix for the memory bug is in imgcodecs.pyx, we need to copy the data
from the np.ndarray as it seems the memory from the img np.ndarray can
be released before the Mat wrapped version is read by `imwrite`.

It turns out you can specify `nogil` at the `cdef extern` level so I've
moved up all the `nogil` specifiers.

If a method isn't specified when invoking the CLI it crashed with an
attribute error, now it prints the help message.