[tests] Remove gstreamer backend test
I'm not keen on keeping this GStreamer test as it requires the opencv
build to have GStreamer, which is yet another dependency to get right.
We already have a test for ffmpeg and one for images verifying the
backend functionality, so I think we can just drop this one as it
doesn't justify its cost.